refactor: modularize project building
Move building logic to its own module file for better organization,
splitting of functions and more unit-testable code.
