+++ src/main.rs
@@ -1,4 +1,4 @@
-use std::{env, fs, path::PathBuf, process::Command};
+use std::{env, process::Command};
 
 use anyhow::{self, Context};
 
@@ -6,6 +6,7 @@ mod cli;
 use cli::{Cli, Commands, Parser};
 
 mod new;
+mod build;
 
 fn main() -> anyhow::Result<()> {
     let cli = Cli::parse();
@@ -19,76 +20,18 @@ fn main() -> anyhow::Result<()> {
         }
         Commands::Build => {
             println!("Building project...");
-            build_project().with_context(|| "Failed to build project.")?;
+            build::build_project().with_context(|| "Failed to build project.")?;
             println!("Project built succesfully!");
         }
         Commands::Run => {
             println!("Building project...");
-            build_project()
+            build::build_project()
                 .with_context(|| "Failed to build project before attempting to run it.")?;
             println!("Project built succesfully!");
             println!("Running project...");
             run_project().with_context(|| "Failed to run project")?;
         }
     };
-
-    Ok(())
-}
-
-fn build_project() -> anyhow::Result<()> {
-    let project_dir = env::current_dir().with_context(|| "Couldn't get project directory.")?;
-    let project_src = project_dir.join("src");
-    let project_target = project_dir.join("target");
-
-    anyhow::ensure!(
-        project_src.exists() && project_target.exists(),
-        format!(
-            "No src {} and target {} directories.",
-            project_src.display(),
-            project_target.display()
-        )
-    );
-
-    let src_files: Vec<PathBuf> = fs::read_dir(&project_src)
-        .with_context(|| format!("Couldn't read source directory {}.", &project_src.display()))?
-        .filter_map(|f| f.ok())
-        .map(|f| f.path())
-        .filter(|f| f.extension().unwrap() == "cpp")
-        .collect();
-
-    anyhow::ensure!(
-        !src_files.is_empty(),
-        format!(
-            "No source `.cpp` files to compile found in \"{}\" directory.",
-            &project_src.display()
-        )
-    );
-
-    let project_name = match project_dir.file_name() {
-        Some(name) => name,
-        None => anyhow::bail!(format!(
-            "Couldn't get project name from {}.",
-            project_dir.display()
-        )),
-    };
-
-    let output_file = project_target.join(project_name);
-
-    let output_args = [
-        "-o",
-        output_file
-            .to_str()
-            .expect("Should be able to convert output_file to string"),
-    ];
-
-    let mut compiler = Command::new("g++");
-    compiler.args(output_args).args(src_files);
-    println!("Running compiler...\n{:?}", &compiler);
-    let compiler_status = compiler
-        .status()
-        .with_context(|| format!("Couldn't start compiler: {:?}", compiler))?;
-
-    anyhow::ensure!(compiler_status.success(), "Compilation failed!");
 
     Ok(())
 }
